knowledge_graph_project/requirements.txt
## Core Dependencies
openai>=1.0.0
anthropic>=0.20.0
google-generativeai>=0.5.0
psycopg2-binary>=2.9.0
neo4j>=5.0.0
pandas>=2.0.0
numpy>=1.24.0

## Machine Learning & NLP
scikit-learn>=1.3.0
networkx>=3.0
transformers>=4.30.0
sentence-transformers>=2.2.0

## Async & API
asyncio
aiohttp>=3.8.0
tenacity>=8.2.0

## Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
click>=8.1.0
pydantic>=2.0.0
loguru>=0.7.0

## Data Processing
jsonlines>=3.1.0
openpyxl>=3.1.0

## Visualization & Monitoring
plotly>=5.15.0
streamlit>=1.28.0

## Development & Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
